Move next link of pointer changes to controller
