+++ content/test/integration/service/deployments/pointer-changes-check.spec.ts
@@ -64,16 +64,6 @@ describe('Integration - Pointer Changes Check', () => {
     assertChangesAre(changes, [P1, { before: E1, after: undefined }], [P2, { before: E2, after: undefined }])
   })
 
-  it('When getting a page of pointer changes then next url is set', async () => {
-    const deploymentTimestamp = await deployEntitiesCombo(service, E1, E3)
-
-    const { pagination: pagination } = await service.getPointerChanges({ entityTypes: [E3.entity.type] }, undefined, 1)
-
-    expect(pagination.next).toContain(`to=${deploymentTimestamp}`)
-    // expect(pagination.next).toContain(`lastId=${E1.entity.id}`)
-    expect(pagination.lastId).toBeUndefined()
-  })
-
   function assertChangesAre(
     changes: PointerChanges,
     ...expectedChanges: [Pointer, { before: EntityCombo | undefined; after: EntityCombo | undefined }][]
